requirements.txt
Brotli==1.0.7
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.10
django-admin-interface==0.12.2
django-colorfield==0.3.2
django-filebrowser-no-grappelli==3.8.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-phone-field==1.8.1
django-tinymce==3.0.2
django-tinymce4-lite==1.8.0
djangorestframework==3.11.0
drf-yasg==1.17.1
gunicorn==20.0.4
idna==2.10
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
jsmin==2.2.2
MarkupSafe==1.1.1
packaging==20.4
Pillow==7.1.1
psycopg2==2.8.5
pyenchant==3.1.1
pyparsing==2.4.7
pytz==2019.3
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
whitenoise==5.1.0
